+++ MatchdayPredictions.Api/Models/Api/User.cs
@@ -5,6 +5,6 @@
         public int UserId { get; init; }
         public string Username { get; init; } = string.Empty;
         public string Email { get; init; } = string.Empty;
-        public string PasswordHash { get; init; } = string.Empty; 
+        public string PasswordHash { get; init; } = string.Empty;
     }
 }