+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+## [0.12.0] - 2025-11-10
+### Added
+- Added model validation.
+
+## [0.11.0] - 2025-11-09
+### Added
+- Added OpenTelemetry metrics and improved JWT authorization flow across the API.
+
 ## [0.10.0] - 2025-11-09
 ### Added
 - Added OpenTelemetry metrics and improved JWT authorization flow across the API.